ok not bad

+++ ENGINE/editor/devtools/room_anchor_tools_panel.hpp
@@ -21,7 +21,7 @@ public:
     };
 
     struct DetailValues {
-        int depth_offset = 0;
+        float depth_offset = 0.0f;
         bool flip_horizontal = true;
         bool flip_vertical = true;
         float rotation_degrees = 0.0f;
